+++ TrelloDotNet/TrelloDotNet/Model/Member.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Text.Json.Serialization;
+using TrelloDotNet.Control;
 
 namespace TrelloDotNet.Model
 {